Split Profile Page into smaller components
